+++ CHANGELOG.md
@@ -7,9 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- accept 'readme' table with 'content-type' and ('file' or 'text')
-- accept 'license' table with 'file' or 'text'
-- accept 'license-file' as a string or list of strings
+- '[project]' section of 'pyproject.toml':
+    - accept 'readme' table with 'content-type' and ('file' or 'text')
+    - accept 'license' table with 'file' or 'text'
+    - accept 'license-file' as a string or list of strings
 
 ## [v0.41.1] - 2023-04-27