+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - accept `license` table with `file` or `text` (this table is deprecated)
     - accept `license-files` as a string, list of strings, or table with `paths` or `globs`
     - default `license-files` globs to `["LICEN[CS]E*", "COPYING*", "NOTICE*", "AUTHORS*"]``
+- rename `cmeel/{helpers -> env}.py`
 
 ## [v0.41.1] - 2023-04-27