+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.42.0] - 2023-05-01
+
 - `[project]` section of `pyproject.toml`:
     - accept `readme` table with `content-type` and (`file` or `text`)
     - accept `license` table with `file` or `text` (this table is deprecated)
@@ -310,6 +312,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial minial working example
 
 [Unreleased]: https://github.com/cmake-wheel/cmeel/compare/v0.4.2...main
+[v0.42.0]: https://github.com/cmake-wheel/cmeel/compare/v0.41.1...v0.42.0
 [v0.41.1]: https://github.com/cmake-wheel/cmeel/compare/v0.41.0...v0.41.1
 [v0.41.0]: https://github.com/cmake-wheel/cmeel/compare/v0.40.0...v0.41.0
 [v0.40.0]: https://github.com/cmake-wheel/cmeel/compare/v0.39.0...v0.40.0